refactor(parse): rename parseAndAdapt to parse
