+++ src/shared/components/HtmlToReactConverter/parse/svgMap.js
